[comp/Login] fix vue empty template warning
