--- src/js/comp/oauth/Login.js
@@ -1,25 +1,0 @@
-import { event } from "../../events.js"
-import Alert from "../Alert.js"
-
-event.init()
-
-export default {
-    mounted() {
-        let promise = window.api.login(this.$route.query["access_token"])
-        promise.then(
-            (account) => {
-                event.emit("account-update")
-                this.$router.push({path: "/"})
-            },
-            (error) => {
-                this.reportError(error)
-            }
-        )
-    },
-
-    mixins: [ Alert ],
-
-    props: {
-        account: { required: true }
-    }
-}
