+++ src/js/main.js
@@ -4,8 +4,8 @@ import VueRouter            from "vue-router"
 import API                  from "./data.js"
 import App                  from "./App.vue"
 import Index                from "./comp/Index.vue"
-import OAuthLogin           from "./comp/oauth/Login.js"
 import Authorize            from "./comp/oauth/Authorize.vue"
+import OAuthLogin           from "./comp/oauth/Login.vue"
 import Search               from "./comp/search/Search.vue"
 import RepoSearch           from "./comp/search/RepoSearch.vue"
 import UserSearch           from "./comp/search/UserSearch.vue"
